+++ src/cli/vars.rs
@@ -3,8 +3,8 @@ use itertools::Itertools;
 
 use super::error::{Error, LibError};
 use super::stringify::stringify;
-use crate::lib::doeval;
-use crate::lib::model::variables::Variable;
+use super::lib::doeval;
+use super::lib::model::variables::Variable;
 
 #[allow(clippy::module_name_repetitions)]
 /// Formats a printable string listing all the `Variables` in the given slice `vars`
